Rename components file
