+++ deeps-engine-latest/CMakeLists.txt
@@ -63,7 +63,7 @@ if(RUN_STANDALONE)
         "${PROJECT_SOURCE_DIR}/src/Engine/include/entt.hpp"
         "${PROJECT_SOURCE_DIR}/src/Engine/Shader.h"
         "${PROJECT_SOURCE_DIR}/src/Engine/Shader.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Engine/component/Components.h"
+        "${PROJECT_SOURCE_DIR}/src/Engine/component/Component.h"
         "${PROJECT_SOURCE_DIR}/src/Engine/Renderer.cpp"
         "${PROJECT_SOURCE_DIR}/src/Engine/Renderer.h"
         "${PROJECT_SOURCE_DIR}/src/Engine/Scene/Scene.cpp"
@@ -86,7 +86,7 @@ else()
         "${PROJECT_SOURCE_DIR}/src/Engine/include/entt.hpp"
         "${PROJECT_SOURCE_DIR}/src/Engine/Shader.h"
         "${PROJECT_SOURCE_DIR}/src/Engine/Shader.cpp"
-        "${PROJECT_SOURCE_DIR}/src/Engine/component/Components.h"
+        "${PROJECT_SOURCE_DIR}/src/Engine/component/Component.h"
         "${PROJECT_SOURCE_DIR}/src/Editor/mainwindow.cpp"
         "${PROJECT_SOURCE_DIR}/src/Editor/OpenGLWidget.h"
         "${PROJECT_SOURCE_DIR}/src/Editor/OpenGLWidget.cpp"
